--- examples/ray-fine-tuning-gpt-j/README.md
@@ -1,4 +1,0 @@
-Fine-tune the GPT-J model:
-```
-python ray_deepspeed_flow.py --environment=conda run
-```
