--- examples/ray-fine-tuning-gpt-j/env.yml
@@ -1,15 +1,0 @@
-name: ray
-channels:
-  - pytorch
-  - conda-forge
-  - defaults
-dependencies:
-  - pip
-  - metaflow
-  - pandas=1.5.3
-  - matplotlib=3.5.3
-  - pip:
-    - opencv_python_headless==4.5.5.62
-    - setuptools==59.5.0
-    - black
-    - python-dotenv
