--- examples/ray-fine-tuning-gpt-j/local.env
@@ -1,1 +1,0 @@
-EN_BATCH=1
